Flooding: Fix memory leak, use uint16_t for bcast_id

+++ elements/brn/routing/broadcast/flooding/flooding.cc
@@ -556,7 +556,7 @@ Flooding::get_broadcast_node(EtherAddress *src)
 }
 
 void
-Flooding::add_id(EtherAddress *src, uint32_t id, Timestamp *now, bool me_src)
+Flooding::add_id(EtherAddress *src, uint16_t id, Timestamp *now, bool me_src)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -570,7 +570,7 @@ Flooding::add_id(EtherAddress *src, uint32_t id, Timestamp *now, bool me_src)
 }
 
 int
-Flooding::add_last_node(EtherAddress *src, uint32_t id, EtherAddress *last_node, bool forwarded, bool rx_acked, bool responsibility, bool foreign_responsibility)
+Flooding::add_last_node(EtherAddress *src, uint16_t id, EtherAddress *last_node, bool forwarded, bool rx_acked, bool responsibility, bool foreign_responsibility)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -583,7 +583,7 @@ Flooding::add_last_node(EtherAddress *src, uint32_t id, EtherAddress *last_node,
 }
 
 void
-Flooding::inc_received(EtherAddress *src, uint32_t id, EtherAddress *last_node)
+Flooding::inc_received(EtherAddress *src, uint16_t id, EtherAddress *last_node)
 {
   uint32_t *cnt = _recv_cnt.findp(*last_node);
   if ( cnt == NULL ) _recv_cnt.insert(*last_node,1);
@@ -600,7 +600,7 @@ Flooding::inc_received(EtherAddress *src, uint32_t id, EtherAddress *last_node)
 }
 
 bool
-Flooding::have_id(EtherAddress *src, uint32_t id, Timestamp *now, uint32_t *count_forwards)
+Flooding::have_id(EtherAddress *src, uint16_t id, Timestamp *now, uint32_t *count_forwards)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -615,7 +615,7 @@ Flooding::have_id(EtherAddress *src, uint32_t id, Timestamp *now, uint32_t *coun
 }
 
 void
-Flooding::forward_attempt(EtherAddress *src, uint32_t id)
+Flooding::forward_attempt(EtherAddress *src, uint16_t id)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -625,7 +625,7 @@ Flooding::forward_attempt(EtherAddress *src, uint32_t id)
 }
 
 void
-Flooding::forward_done(EtherAddress *src, uint32_t id, bool success, bool new_node)
+Flooding::forward_done(EtherAddress *src, uint16_t id, bool success, bool new_node)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -637,7 +637,7 @@ Flooding::forward_done(EtherAddress *src, uint32_t id, bool success, bool new_no
 }
 
 uint32_t
-Flooding::unfinished_forward_attempts(EtherAddress *src, uint32_t id)
+Flooding::unfinished_forward_attempts(EtherAddress *src, uint16_t id)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -647,7 +647,7 @@ Flooding::unfinished_forward_attempts(EtherAddress *src, uint32_t id)
 }
 
 void
-Flooding::sent(EtherAddress *src, uint32_t id, uint32_t no_transmission)
+Flooding::sent(EtherAddress *src, uint16_t id, uint32_t no_transmission)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -657,7 +657,7 @@ Flooding::sent(EtherAddress *src, uint32_t id, uint32_t no_transmission)
 }
 
 bool
-Flooding::me_src(EtherAddress *src, uint32_t id)
+Flooding::me_src(EtherAddress *src, uint16_t id)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -715,7 +715,7 @@ Flooding::reset()
 }
 
 struct Flooding::BroadcastNode::flooding_last_node*
-Flooding::get_last_node(EtherAddress *src, uint32_t id, EtherAddress *last)
+Flooding::get_last_node(EtherAddress *src, uint16_t id, EtherAddress *last)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
   if ( bcn == NULL ) return NULL;
@@ -723,7 +723,7 @@ Flooding::get_last_node(EtherAddress *src, uint32_t id, EtherAddress *last)
 }
 
 struct Flooding::BroadcastNode::flooding_last_node*
-Flooding::get_last_nodes(EtherAddress *src, uint32_t id, uint32_t *size)
+Flooding::get_last_nodes(EtherAddress *src, uint16_t id, uint32_t *size)
 {
   *size = 0;
   BroadcastNode *bcn = _bcast_map.find(*src);
@@ -732,7 +732,7 @@ Flooding::get_last_nodes(EtherAddress *src, uint32_t id, uint32_t *size)
 }
 
 void
-Flooding::assign_last_node(EtherAddress *src, uint32_t id, EtherAddress *last_node)
+Flooding::assign_last_node(EtherAddress *src, uint16_t id, EtherAddress *last_node)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
 
@@ -743,7 +743,7 @@ Flooding::assign_last_node(EtherAddress *src, uint32_t id, EtherAddress *last_no
 }
 
 struct Flooding::BroadcastNode::flooding_last_node*
-Flooding::get_assigned_nodes(EtherAddress *src, uint32_t id, uint32_t *size)
+Flooding::get_assigned_nodes(EtherAddress *src, uint16_t id, uint32_t *size)
 {
   *size = 0;
   BroadcastNode *bcn = _bcast_map.find(*src);
@@ -752,28 +752,28 @@ Flooding::get_assigned_nodes(EtherAddress *src, uint32_t id, uint32_t *size)
 }
 
 void
-Flooding::clear_assigned_nodes(EtherAddress *src, uint32_t id)
+Flooding::clear_assigned_nodes(EtherAddress *src, uint16_t id)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
   if ( bcn != NULL ) return bcn->clear_assigned_nodes(id);  
 }
 
 void
-Flooding::set_responsibility_target(EtherAddress *src, uint32_t id, EtherAddress *target)
+Flooding::set_responsibility_target(EtherAddress *src, uint16_t id, EtherAddress *target)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
   if ( bcn != NULL ) bcn->set_responsibility_target(id, target);
 }
 
 void
-Flooding::set_foreign_responsibility_target(EtherAddress *src, uint32_t id, EtherAddress *target)
+Flooding::set_foreign_responsibility_target(EtherAddress *src, uint16_t id, EtherAddress *target)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
   if ( bcn != NULL ) bcn->set_foreign_responsibility_target(id, target);
 }
 
 bool
-Flooding::is_responsibility_target(EtherAddress *src, uint32_t id, EtherAddress *target)
+Flooding::is_responsibility_target(EtherAddress *src, uint16_t id, EtherAddress *target)
 {
   BroadcastNode *bcn = _bcast_map.find(*src);
   if ( bcn != NULL ) return bcn->is_responsibility_target(id, target);
